BUG: Fix DFS visitor edge_points in tree_edge
When starting on nodes with high degree, even after hitting
finish_on_junction, it wouldn't stop the unwinding of the stack
for the rest of the branches. This is by design of DFS algorithm.

We now clear in tree_edge the number of edge_points in current edge.
To avoid the bug of having edge_points from an adjacent edge in other
edge.