Wrap view_spatial_graph and view_spatial_graph_with_points
Sadly, it shows a bug in VTK, or an erroneous handling on sgext side
in the `view_spatial_graph_with_points`. When running it in ipython
the window does not close when clicked in the `X` icon. Instead, the
window becomes an unresponsive static image, and the only way to
properly close it is shutting down `ipython`.

Similar error was reported in VTK a couple years ago.
I reported there this case:

https://gitlab.kitware.com/vtk/vtk/-/issues/17251#note_754293