Add tables_folder argument to thin functions.
`tables_folder` is a string holding the path to the folder containing
look-up-tables used in DGtal for thinning.
- simplicity_table26_6.zlib
- isthmusicityOne_table26_6.zlib
- isthmusicity_table26_6.zlib

Before this patch, `thin_function` was using the hardcoded table path
from the generated DGtal header `<DGtal/topology/tables/NeighborhoodTables.h>`
The paths in that header are correct when DGtal has been built and/or
installed in the same system where you execute the thinning. However,
the paths are obviously not correct when distributing the library to
other machines.

The `cpp-scripts/thin` accepts a tables_folder argument, but use as
default the one from DGtal, so nothing to change for users.

In python, the 3 aforementioned look-up tables are distributed along the
sgext package (using `package_data` from setuptools) and can be found in
`sgext/tables` folder.
To facilitate its use, a variable tables_folder is populated in the
`__init__` to the location of that folder.
```python
from sgext import tables_folder
import sgext
tables_folder=sgext.tables_folder
thin(...
    tables_folder=tables_folder
    ...)
```