Add collapse_clusters vistitor and high level interface function
Allows using vertex_descriptors near by the centroids of clusters,
so the edge points of edges going out of a cluster are still correct.

If the user is interested in the real centroid, we would need to move the
position of the spatial node and also generate edge points from that
new position.
This would require some kind of voxelization of a line.
VTK might have something, but usually not really good with 1D lines. However
somekind of triangulation/mesh might also work.

See:
http://vtk.1045678.n5.nabble.com/Solid-Voxelization-with-VTK-td5747520.html