Thin: Avoid creating a DGtal Container when importing a distance map
Adds a function `SG::select_max_value_of_clique` that works
directly with ITK images.

Before, we were using the DGtal function selectMaxValue that needs
a DGtal type.

This should save a good amount of memory, but more importantly, it
allows to use distance maps computed from ITK, that takes into account
spacing and direction.
In DGtal the distance maps doesn't take into account the metadata,
and the greatest distance always corresponds to the central voxel
of the vessel. This might lead to incorrect centers, specially in
images with non-isotropic spacings.