Visualize: Add view_image_with_label function
This function visualize any single component input image along a
label/binarized image of the same extent.

The visualization show both image superimposed (blended),
and the cursor interactor show the value of the original images.

The purpose is to check how good is an automatic binarization against
the original image.

Adds `sgextImagePlaneWidget` deriving from `vtkImagePlaneWidget`.
The input of the widget would be a blend images of multiple images.
This class allows to display the values of the original images before the blend.
through the function: `AddImageForCursorDiplay`.

Add python wrappings: `view_image_with_label`