WIP: Add emscripten support
Use emcmake to use the emscripten toolchain

There is a problem on find_package with the toolchain,
not sure where, but seems related with CMAKE_PREFIX_PATH or similar.

This commit show the problem, `find_package(Boost)` with COMPONENTS
do not find the components. The components use HINTS based on the
Boost_DIR, but the HINTS fails...