CMake: Fix `export(TARGETS` to the build tree
The file SGEXTTargets.cmake in the build tree was huge, because the APPEND
option used.

Instead, all the targets are collected in the
(`PARENT_SCOPE`) `_targets_to_export` variable.
And the exported at once without the APPEND option.

Also add the target `pybind11_headers` to fix CMake error
that pybind11::module depends on it, but wasn't exported.
This pybind11 seems related to changes (good!) on pybind11 CMake.