SpatialNode initializes pos using value initialization instead of default
Instead of the std::array pos being default initialzed (whatever doubles)
force it to use value initialization, so the pos are initialized to
pos(0,0,0). Mainly for sanity.

This showed a bug in  with the fixture `sg_extra_connected_junction`

It had wrong number of vertices. The associated fixed test is the following:

```bash
 ctest -R Extract__test_spatial_graph_reduction__extra_connected_junctions.reduce  -VV
```